# docs/swagger.yaml
basePath: /
definitions:
  handlers.IDsRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  models.GrantMemoryNodeAccessRequest:
    properties:
      userId:
        type: integer
    required:
    - userId
    type: object
...
  models.MemoryNodeFull:
    properties:
      aliases:
        items:
          type: string
        type: array
      cards:
        items:
          type: integer
        type: array
      children:
        items:
          type: integer
        type: array
      description:
        type: string
      groups:
        items:
          $ref: '#/definitions/schema.CardsGroup'
        type: array
      id:
        type: integer
      name:
        type: string
      parents:
        items:
          type: integer
        type: array
      priorities:
        items:
          $ref: '#/definitions/schema.CardsPriority'
        type: array
      shared:
        type: boolean
      userId:
        type: integer
    type: object
...
  models.MemoryNodePathItem:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.MoveSharedNodeToUserRequest:
    properties:
      deep:
        type: boolean
      userId:
        type: integer
    required:
    - userId
    type: object
...
  models.NewMemoryNodeRequest:
    properties:
      memoryNode:
        $ref: '#/definitions/models.MemoryNodeShort'
    required:
    - memoryNode
    type: object
...
  models.RemoveSharedNodeFromUserRequest:
    properties:
      deep:
        type: boolean
      userId:
        type: integer
    required:
    - userId
    type: object
...
host: localhost:3033
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Memory nodes, cards, and card items API server.
    **Authentication:** Click **Authorize**, enter your email as username and your password, then Authorize. Swagger will call `POST /users/token` and attach the access token as a Bearer header.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: MemoryGuard API
  version: "1.0"
paths:
...
  /admin/memory-node/{id}/move-to-user:
    post:
      consumes:
      - application/json
      description: Admin-only. Grants a shared memory node and its shared cards to
        a user. When deep=true, also grants shared descendants recursively.
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      - description: Target user and deep flag
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MoveSharedNodeToUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Move shared memory node to user (admin)
      tags:
      - memory-nodes
  /admin/memory-node/{id}/remove-from-user:
    post:
      consumes:
      - application/json
      description: Admin-only. Revokes a user's access by deleting memory_node_users
        / card_users (and card_user_counts) links. Does not delete the node or cards.
        When deep=true, also revokes shared descendants recursively.
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      - description: Target user and deep flag
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RemoveSharedNodeFromUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Remove shared memory node from user (admin)
      tags:
      - memory-nodes
...
  /get-memory-nodes:
    post:
      consumes:
      - application/json
      description: Returns multiple memory nodes by their IDs for the authenticated
        user
      parameters:
      - description: List of memory node IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MemoryNodeFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get memory nodes by IDs
      tags:
      - memory-nodes
...
  /memory-node-by-alias/{alias}:
    get:
      description: Finds a memory node by one of its aliases for the authenticated
        user
      parameters:
      - description: Alias
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MemoryNodeFull'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get memory node by alias
      tags:
      - memory-nodes
  /memory-node/{id}:
    delete:
      description: Deletes a memory node owned by the authenticated user. If the node
        is shared, also deletes owned child nodes, their cards, and all memory_node_users
        / card_users (and card_user_counts) links.
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Delete memory node
      tags:
      - memory-nodes
    get:
      description: Returns a memory node by its ID if the authenticated user owns
        it or has an explicit shared grant
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MemoryNodeFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get memory node by ID
      tags:
      - memory-nodes
...
  /memory-node/{id}/parents-path:
    get:
      description: Returns the first-parent chain from root to the given node (inclusive)
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MemoryNodePathItem'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Get parents path for a memory node
      tags:
      - memory-nodes
  /memory-node/{id}/users:
    post:
      consumes:
      - application/json
      description: Creates a memory_node_users row (and card_users for shared cards
        on the node) so the user can see the shared node. Does not make the node visible
        to everyone.
      parameters:
      - description: Memory node ID
        in: path
        name: id
        required: true
        type: integer
      - description: User to grant
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GrantMemoryNodeAccessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Grant shared memory node access
      tags:
      - memory-nodes
  /memory-nodes:
    get:
      description: Returns all memory nodes for the authenticated user, or a subset
        when ids query is provided as a JSON array
      parameters:
      - description: JSON array of IDs, e.g. [1,2,3]
        in: query
        name: ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MemoryNodeFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: List memory nodes
      tags:
      - memory-nodes
  /memory-nodes/roots:
    get:
      description: Returns memory nodes with no parents for the authenticated user
        (owned or explicitly granted shared roots)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MemoryNodeFull'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: List root memory nodes
      tags:
      - memory-nodes
...
  /new-memory-node:
    post:
      consumes:
      - application/json
      description: Creates a new memory node owned by the authenticated user
      parameters:
      - description: Memory node creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.NewMemoryNodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MemoryNodeFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Create memory node
      tags:
      - memory-nodes
...
  /update-memory-node:
    put:
      consumes:
      - application/json
      description: Partially updates an existing memory node owned by the authenticated
        user
      parameters:
      - description: Memory node update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MemoryNodePartial'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MemoryNodeFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Login:
        - api
      summary: Update memory node
      tags:
      - memory-nodes
...
schemes:
- http
- https
security:
- Login:
  - api
securityDefinitions:
  Login:
    flow: password
    scopes:
      api: Access the MemoryGuard API
    tokenUrl: /users/token
    type: oauth2
swagger: "2.0"